+++ src/main/kotlin/com/dsoftware/ghmanager/ui/panels/filters/WfRunsSearchPanelViewModel.kt
@@ -22,7 +22,8 @@ internal class WfRunsSearchPanelViewModel(
         get() = context.runsListLoader.repoBranches
     val collaborators
         get() = context.runsListLoader.repoCollaborators
-
+    val workflowTypes
+        get() = context.runsListLoader.workflowTypes
     override fun WfRunsListSearchValue.withQuery(query: String?) = copy(searchQuery = query)
 
     override val quickFilters: List<WorkflowRunListQuickFilter> = listOf(
@@ -33,12 +34,12 @@ internal class WfRunsSearchPanelViewModel(
     val eventFilterState = searchState.partialState(WfRunsListSearchValue::event) { copy(event = it) }
     val userFilterState = searchState.partialState(WfRunsListSearchValue::actor) { copy(actor = it) }
     val statusState = searchState.partialState(WfRunsListSearchValue::status) { copy(status = it) }
-
+    val workflowType = searchState.partialState(WfRunsListSearchValue::workflowType) { copy(workflowType = it) }
 }
 
 sealed class WorkflowRunListQuickFilter(val title: String) : ReviewListQuickFilter<WfRunsListSearchValue> {
 
-    class All : WorkflowRunListQuickFilter("All workflows") {
+    class All : WorkflowRunListQuickFilter("All workflow runs") {
         override val filter = WfRunsListSearchValue()
     }
 }