+++ CHANGELOG.md
@@ -4,9 +4,19 @@
 
 ## [Unreleased]
 
+## [1.14.0]
+
+### 🚀 Features
+
+- Filter by workflow type #98
+
+### Maintenance
+
+- Improve filter behavior
+
 ## [1.13.5]
 
-### Maintainance
+### Maintenance
 
 - Add support for build 223.3 and fixed a few warnings. @wyatt-herkamp #97